#48 moved states commands to a separate trigger package and renamed them
